backend/requirements.txt
# -----------------------------------------
# Core dependencies
# -----------------------------------------
python-slugify==8.0.4
Pillow==11.1.0
argon2-cffi==23.1.0

# -----------------------------------------
# Redis and caching
# -----------------------------------------
redis==5.2.1
hiredis==3.1.0

# -----------------------------------------
# Task processing
# -----------------------------------------
celery==5.5.0
django-celery-beat==2.7.0
django-celery-email==3.0.0
flower==2.0.1

# -----------------------------------------
# Web server
# -----------------------------------------
uvicorn[standard]==0.34.0
uvicorn-worker==0.3.0
gunicorn==23.0.0

# -----------------------------------------
# Django core
# -----------------------------------------
django==5.0.13
django-environ==0.12.0
django-model-utils==5.0.0
django-redis==5.4.0

# -----------------------------------------
# Django REST Framework
# -----------------------------------------
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework-simplejwt==5.5.0

# -----------------------------------------
# API documentation
# -----------------------------------------
drf-spectacular==0.28.0
drf-spectacular[sidecar]==0.28.0

# -----------------------------------------
# Database
# -----------------------------------------
psycopg==3.2.6

# -----------------------------------------
# Storage
# -----------------------------------------
Collectfasta==3.2.1
django-storages[s3]==1.14.5

# -----------------------------------------
# Monitoring
# -----------------------------------------
sentry-sdk==2.25.0
django-silk==5.3.2

# -----------------------------------------
# Development tools
# -----------------------------------------
Werkzeug[watchdog]==3.1.3
ipdb==0.13.13
watchfiles==1.0.4

# -----------------------------------------
# Code quality
# -----------------------------------------
ruff==0.11.2
djlint==1.36.4
pre-commit==4.2.0

# -----------------------------------------
# Django extensions
# -----------------------------------------
django-extensions==3.2.3
django-health-check==3.18.3

# -----------------------------------------
# Agent frameworks
# -----------------------------------------
ag2==0.8.5
ag2[gemini]==0.8.5
ag2[ollama]==0.8.5

# -----------------------------------------
# Secret management
# -----------------------------------------
hvac==2.3.0

